Move data calculations to chartUtils
